refactor: move experimental setpoint types under 'experimental'
These will likely still change.
