+++ examples/cpp/modes/rtl_replacement/include/mode.hpp
@@ -7,7 +7,7 @@
 #include <px4_ros2/components/mode.hpp>
 #include <px4_ros2/components/health_and_arming_checks.hpp>
 #include <px4_ros2/components/mode_executor.hpp>
-#include <px4_ros2/control/setpoint_types/trajectory.hpp>
+#include <px4_ros2/control/setpoint_types/experimental/trajectory.hpp>
 #include <px4_msgs/msg/vehicle_land_detected.hpp>
 
 #include <rclcpp/rclcpp.hpp>
