+++ px4_ros2_cpp/src/control/setpoint_types/experimental/attitude.cpp
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  ****************************************************************************/
 
-#include <px4_ros2/control/setpoint_types/attitude.hpp>
+#include <px4_ros2/control/setpoint_types/experimental/attitude.hpp>
 
 
 namespace px4_ros2
